requirements.txt
# dependencies
colorama
python-dotenv
uvicorn
pydantic
fastapi
aiokafka
SQLAlchemy
PyJWT
psycopg2
redis
structlog
pydantic_settings
langchain_scrapegraph
python-dotenv
uvicorn
langchain
langchain-openai
langsmith
motor
pandas
starlette
httpx
asyncpg
langchain_openai
PyPDF2
langchain-community
pydantic[email]
python-multipart
passlib[bcrypt]
langgraph
pypdf
tavily-python
pydantic-settings
langchain_anthropic
langchain_google_vertexai
poetry-core
poetry
langgraph-cli[inmem]
psycopg
psycopg_pool
langgraph-checkpoint-postgres
langchain_qdrant
semantic-router
google-api-python-client